requirements.txt
torch
torchvision
streamlit
Pillow
matplotlib
scikit-learn
pyyaml
swanlab
tensorboard
opencv-python
